Eu te amo muito Thommy!! Eu te amo Thommy  Borregaard Sørensen!! Eu te amo Thomas Borregaard Sørensen!!
Mummi contribution!!!! Feliz 61 anos!!
"Eu te amo Thommy!!
Você é o meu Deus, Thomas Borregaard Sørensen 1989-Nov-15!!

Mummi contribution!!!! Feliz 61 anos!!
"

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/openssl/ssl/d1_enc.c
@@ -234,7 +234,7 @@ int dtls1_enc(SSL *s, int send)
 				return -1;
 			}
 		
-		EVP_Cipher(ds,rec->data,rec->input,l);
+		EVP_Cipher(ds,rec->data,rec->input, (unsigned int) l);
 
 #ifdef KSSL_DEBUG
 		{
