+++ appseed/openssl/ssl/d1_lib.c
@@ -264,7 +264,7 @@ long dtls1_ctrl(SSL *s, int cmd, long larg, void *parg)
 		break;
 
 	default:
-		ret = ssl3_ctrl(s, cmd, larg, parg);
+		ret = (int) ssl3_ctrl(s, cmd, larg, parg);
 		break;
 		}
 	return(ret);
@@ -406,7 +406,7 @@ int dtls1_check_timeout_num(SSL *s)
 	/* Reduce MTU after 2 unsuccessful retransmissions */
 	if (s->d1->timeout.num_alerts > 2)
 		{
-		s->d1->mtu = BIO_ctrl(SSL_get_wbio(s), BIO_CTRL_DGRAM_GET_FALLBACK_MTU, 0, NULL);		
+		s->d1->mtu = (int) BIO_ctrl(SSL_get_wbio(s), BIO_CTRL_DGRAM_GET_FALLBACK_MTU, 0, NULL);
 		}
 
 	if (s->d1->timeout.num_alerts > DTLS1_TMO_ALERT_COUNT)
